Deploy kube-flannel as a ClusterResourceSet (#82)
Move the kube-flannel manifest outside of the clusterclass
(part of the KubeadmConfig) and into the cluster templates
(deployed as a ClusterResourceSet).

This allows upgrading the kube-flannel manifests without
causing a rollout restart in existing cluster machines because
the KubeadmConfigTemplate is different.

Also, use the kube-flannel from the cluster template for the
e2e tests, rather than maintaining a separate kube-flannel
manifest.

---------

Signed-off-by: Angelos Kolaitis <neoaggelos@gmail.com>